refactor(app): move `UsersPage` component under the `users` directory
